Add support for unit test in Rust code
Rust code depends on a few symbols exported from octopus. This is not a
problem, as long as rust linked as a library to the main binary.
However, if you try to run unit test then nothing works, because those
dependencies are missing.

Unfortunately, I wasn't able to find a way to provide link deps for a
dev target. Therefore, dummy crate "test_runtime" is used as a workaround.
The only purpose of "test_runtime" crate is to force linking with
octopus.a and the other libs.

rustc linker can't link OCaml runtime, so if you want to run unit tests,
OCaml must be disabled in ./configure